remove old log
