--- packages/localnet/src/log.ts
@@ -1,10 +1,0 @@
-import ansis from "ansis";
-
-export const chains: Record<string, { color: any; name: string }> = {
-  "103": { color: ansis.blue, name: "Sui" },
-  "2015141": { color: ansis.blueBright, name: "TON" },
-  "5": { color: ansis.cyan, name: "Ethereum" },
-  "7001": { color: ansis.green, name: "ZetaChain" },
-  "901": { color: ansis.magenta, name: "Solana" },
-  "97": { color: ansis.yellow, name: "BNB" },
-};
